requirements.txt
python-telegram-bot==21.3
openai==1.30.0
instagrapi==2.0.0
schedule==1.2.1
